Add a new way to specify dialogue templates
Remove transition templates, and replace them with separate
$agent and $user templates.

This simplifies how semantic functions are written, as it separates
the computation of the agent and the user state. It also reduces
code duplication, as the same $agent templates are used at
inference time.

The challenge in this design is how information is passed from
the agent utterance to the user utterance, when that information
is not part of the ThingTalk agent state. This is now solved
by having the agent semantic function compute the new context
info, bypassing the context function and adding any information
as necessary.

To further simplify the templates, all semantic functions implementing
the dialogue acts are moved to a new package, with modules
named the same was the corresponding template files.
