+++ languages/thingtalk/policy.yaml
@@ -1,0 +1,59 @@
+name: 'org.thingpedia.dialogue.transaction'
+
+dialogueActs:
+  user:
+    # user says hi!
+    - greet
+    # user issues a ThingTalk program
+    - execute
+    # user wants to see the result of the previous program (in reply to a generic search question)
+    - ask_recommend
+    # user insists in reiterating the same search after an empty search error
+    - insist
+    # user wants to see more output from the previous result
+    - learn_more
+    # user asks to see an output parameter from the previous result
+    - action_question
+    # user says closes the dialogue mid-way (in the middle of a search)
+    - cancel
+    # user terminates the dialogue after the agent asked if there is anything
+    # else the user wants
+    # "end" is a terminal state, it has no continuations
+    # (the agent replies with sys_goodbye which itself generates no user reply)
+    - end
+  agent:
+    # agent says hi back
+    - sys_greet
+    # agent asks a question to refine a query (with or without a parameter)
+    - sys_search_question
+    - sys_generic_search_question
+    # agent asks a question to slot fill a program
+    - sys_slot_fill
+    # agent recommends one, two, or three results from the program (with or without an action)
+    - sys_recommend_one
+    - sys_recommend_two
+    - sys_recommend_three
+    # agent proposes a refined query
+    - sys_propose_refined_query
+    # agent asks the user what they would like to hear
+    - sys_learn_more_what
+    # agent informs that the search is empty (with and without a slot-fill question)
+    - sys_empty_search_question
+    - sys_empty_search
+    # agent executed the action successfully (and shows the result of the action)
+    - sys_action_success
+    # agent had an error in executing the action (with and without a slot-fill question)
+    - sys_action_error_question
+    - sys_action_error
+    # agent asks if anything else is needed
+    - sys_anything_else
+    # agent says good bye
+    - sys_end
+  withParam:
+    - action_question
+    - sys_search_question
+    - sys_slot_fill
+    - sys_empty_search_question
+    - sys_action_error_question
+
+terminalAct: sys_end
