+++ lib/languages/dlgthingtalk.js
@@ -236,6 +236,21 @@ module.exports = {
         return [code, entities];
     },
 
+    validateState(state, forTarget) {
+        if (forTarget === 'user') {
+            // check that there are no 'proposed' items
+            // (should be executed, 'accepted' or 'confirmed')
+            for (let item of state.history)
+                assert(item.confirm !== 'proposed');
+        } else {
+            // check that there are no 'confirmed' items that were not executed
+            // TODO: if we add "intermediate_context" capabilities to the state machine
+            // we can relax this restriction
+            for (let item of state.history)
+                assert(item.confirm !== 'confirmed' || item.results !== null);
+        }
+    },
+
     computeNewState(oldState, prediction, forTarget) {
         return computeNewState(oldState, prediction, forTarget);
     },
