+++ lib/sentence-generator/compiler/grammar.js
@@ -179,8 +179,8 @@ function peg$parse(input, options) {
       peg$c28 = function(first, rest) {
           return new Ast.Statement.Context([first].concat(take(rest, 2)));
       },
-      peg$c29 = function(pname, body) {
-          return new Ast.Statement.ContextTagger(pname, body);
+      peg$c29 = function(name, pname, body) {
+          return new Ast.Statement.FunctionDeclaration(name[0], pname, body);
       },
       peg$c30 = function(body) { return take(body, 0); },
       peg$c31 = function(body) { return [body]; },
@@ -740,7 +740,7 @@ function peg$parse(input, options) {
     if (s0 === peg$FAILED) {
       s0 = peg$parseContextDeclaration();
       if (s0 === peg$FAILED) {
-        s0 = peg$parseContextTaggerDeclaration();
+        s0 = peg$parseFunctionDeclaration();
         if (s0 === peg$FAILED) {
           s0 = peg$parseForLoop();
           if (s0 === peg$FAILED) {
@@ -1109,15 +1109,18 @@ function peg$parse(input, options) {
     return s0;
   }
 
-  function peg$parseContextTaggerDeclaration() {
+  function peg$parseFunctionDeclaration() {
     var s0, s1, s2, s3, s4, s5, s6, s7, s8, s9, s10, s11, s12, s13;
 
     s0 = peg$currPos;
-    s1 = peg$parseContextToken();
+    s1 = peg$parseFunctionToken();
     if (s1 !== peg$FAILED) {
       s2 = peg$parse__();
       if (s2 !== peg$FAILED) {
-        s3 = peg$parseFunctionToken();
+        s3 = peg$parseContextToken();
+        if (s3 === peg$FAILED) {
+          s3 = peg$parseIdentifier();
+        }
         if (s3 !== peg$FAILED) {
           s4 = peg$parse__();
           if (s4 !== peg$FAILED) {
@@ -1164,7 +1167,7 @@ function peg$parse(input, options) {
                             }
                             if (s13 !== peg$FAILED) {
                               peg$savedPos = s0;
-                              s1 = peg$c29(s7, s12);
+                              s1 = peg$c29(s3, s7, s12);
                               s0 = s1;
                             } else {
                               peg$currPos = s0;
