+++ lib/sentence-generator/compiler/grammar.pegjs
@@ -56,7 +56,7 @@ CodeBlock = '{' code:Code '}' { return new Ast.Statement.CodeBlock(code); }
 Statement
   = NonTerminalDeclaration
   / ContextDeclaration
-  / ContextTaggerDeclaration
+  / FunctionDeclaration
   / ForLoop
   / IfStmt
   / ImportStmt
@@ -81,8 +81,8 @@ ContextDeclaration = ContextToken __ first:Identifier rest:(',' __ Identifier __
     return new Ast.Statement.Context([first].concat(take(rest, 2)));
 }
 
-ContextTaggerDeclaration = ContextToken __ FunctionToken __ '(' __ pname:Identifier __ ')' __ '{' body:Code '}' {
-    return new Ast.Statement.ContextTagger(pname, body);
+FunctionDeclaration = FunctionToken __ name:(ContextToken / Identifier) __ '(' __ pname:Identifier __ ')' __ '{' body:Code '}' {
+    return new Ast.Statement.FunctionDeclaration(name[0], pname, body);
 }
 
 StatementOrBlock
