+++ lib/sentence-generator/compiler/meta_ast.js
@@ -87,9 +87,6 @@ class NonTerminalStmt extends Statement {
 
         this.isNonTerminal = true;
 
-        // $user is the same as $root, and it's introduced for symmetry with $agent
-        if (name === '$user')
-            name = '$root';
         this.name = name;
         this.rules = rules;
     }
@@ -118,20 +115,21 @@ class ContextStmt extends Statement {
 }
 Statement.Context = ContextStmt;
 
-class ContextTaggerStmt extends Statement {
-    constructor(pname, code) {
+class FunctionDeclarationStmt extends Statement {
+    constructor(name, pname, code) {
         super();
 
         this.isContextTagger = true;
+        this.name = name;
         this.pname = pname;
         this.code = code;
     }
 
     codegen(prefix = '') {
-        return `${prefix}$grammar.setContextTagger((${this.pname}) => {${this.code}});\n`;
+        return `${prefix}$grammar.declareFunction('${this.name}', (${this.pname}) => {${this.code}});\n`;
     }
 }
-Statement.ContextTagger = ContextTaggerStmt;
+Statement.FunctionDeclaration = FunctionDeclarationStmt;
 
 class ForLoop extends Statement {
     constructor(head, statements) {
