+++ lib/sentence-generator/index.js
@@ -12,8 +12,9 @@
 const assert = require('assert');
 const stream = require('stream');
 const events = require('events');
-
-const { parseEntity } = require('../utils/misc-utils');
+const yaml = require('js-yaml');
+const util = require('util');
+const fs = require('fs');
 
 const $runtime = require('./runtime');
 const i18n = require('../i18n');
@@ -31,17 +32,12 @@ class SentenceGenerator extends events.EventEmitter {
         this._templateFiles = options.templateFiles;
         this._options = options;
         this._langPack = i18n.get(options.locale);
-        this._postprocess = this._langPack.postprocessSynthetic;
         this._grammar = null;
         this._generator = null;
     }
 
     get progress() {
         return this._grammar.progress;
-    }
-
-    postprocess(sentence, program) {
-        return this._postprocess(sentence, program, this._options.rng);
     }
 
     generate(contexts, callback) {
@@ -57,6 +53,7 @@ class SentenceGenerator extends events.EventEmitter {
             const compiledTemplate = await importGenie(file);
             this._grammar = await compiledTemplate(this._options, this._langPack, this._grammar);
         }
+        this._grammar.finalize();
         this._grammar.on('progress', (value) => {
             this.emit('progress', value);
         });
@@ -71,6 +68,8 @@ class BasicSentenceGenerator extends stream.Readable {
             options.targetPruningSize = BASIC_TARGET_GEN_SIZE;
         this._target = require('../languages/' + options.targetLanguage);
         this._idPrefix = options.idPrefix || '';
+        this._langPack = i18n.get(options.locale);
+        this._rng = options.rng;
         this._generator = new SentenceGenerator(options);
         this._generator.on('progress', (value) => {
             this.emit('progress', value);
@@ -95,12 +94,17 @@ class BasicSentenceGenerator extends stream.Readable {
         });
     }
 
+    _postprocessSentence(derivation, program) {
+        let utterance = derivation.toString();
+        utterance = utterance.replace(/ +/g, ' ');
+        utterance = this._langPack.postprocessSynthetic(utterance, program, this._rng);
+        return utterance;
+    }
+
     _output(depth, derivation) {
         let program = derivation.value.optimize();
         assert(program !== null); // not-null even after optimize
-        let preprocessed = derivation.toString();
-        preprocessed = preprocessed.replace(/ +/g, ' ');
-        preprocessed = this._generator.postprocess(preprocessed, program);
+        const preprocessed = this._postprocessSentence(derivation, program);
         let sequence;
         try {
             sequence = this._target.serialize(program, [], {});
@@ -126,22 +130,10 @@ class BasicSentenceGenerator extends stream.Readable {
 }
 
 class PartialDialogue {
-    constructor(context = null, turns = [], constants = {}, execState = undefined) {
+    constructor(context = null, turns = [], execState = undefined) {
         this.context = context;
         this.turns = turns;
-        this.constants = constants;
-        this.execState = undefined;
-    }
-}
-
-function mergeConstants(constants, sentence) {
-    for (let token of sentence.split(' ')) {
-        const entity = parseEntity(token);
-        if (entity === null)
-            continue;
-
-        const [type, index] = entity;
-        constants[type] = Math.max(constants[type] || 0, index + 1);
+        this.execState = execState;
     }
 }
 
@@ -153,10 +145,13 @@ const FACTORS = [50, 75, 75, 100];
  * This object is created afresh for every minibatch.
  */
 class MinibatchDialogueGenerator {
-    constructor(sentenceGenerator, targetLanguage, dialogueAgent, options, minibatchIdx) {
-        this._sentenceGenerator = sentenceGenerator;
+    constructor(agentGenerator, userGenerator, targetLanguage, langPack, dialogueAgent, stateValidator, options, minibatchIdx) {
+        this._agentGenerator = agentGenerator;
+        this._userGenerator = userGenerator;
         this._target = targetLanguage;
+        this._langPack = langPack;
         this._dialogueAgent = dialogueAgent;
+        this._stateValidator = stateValidator;
         this._minibatchSize = options.minibatchSize;
         this._rng = options.rng;
         this._options = options;
@@ -188,69 +183,95 @@ class MinibatchDialogueGenerator {
         return discarded !== dlg;
     }
 
-    _addTurn(continuations, derivation) {
-        const dlg = derivation.context.dlg;
+    _postprocessSentence(derivation, program) {
+        let utterance = derivation.toString();
+        utterance = utterance.replace(/ +/g, ' ');
+        utterance = this._langPack.postprocessSynthetic(utterance, program, this._rng);
+        return utterance;
+    }
 
-        let [agentState, userState] = derivation.value;
-        let agent, agent_target, user, user_target;
-        let newConstants = {};
-        Object.assign(newConstants, dlg.constants);
-        if (dlg.context === null) {
-            // first turn in the dialogue
-            assert(dlg === this._emptyDialogue);
-            assert(agentState === null);
-            userState = userState.optimize();
-            assert(userState !== null); // not-null even after optimize
-            let turnString = derivation.toString();
-            turnString = turnString.replace(/ +/g, ' ');
-            turnString = this._sentenceGenerator.postprocess(turnString, userState);
-            mergeConstants(newConstants, turnString);
+    _generateAgent(partials) {
+        const agentTurns = [];
+        this._agentGenerator.generate(partials, (depth, derivation) => {
+            // derivation.dlg is the PartialDialogue that is being continued
+            // derivation.value is the object returned by the root semantic function, with:
+            // - state (the thingtalk state)
+            // - context (the context info to pass to the semantic function of the user)
+            // - tags (context tags to set when generating user sentences)
+            // - other properties only relevant to inference time we don't care about
 
-            agent = '';
-            agent_target = null;
-            user = turnString;
-            user_target = userState.prettyprint();
-        } else {
-            agentState = agentState.optimize();
-            userState = userState.optimize();
-            assert(agentState !== null); // not-null even after optimize
-            assert(userState !== null); // not-null even after optimize
-            let turnString = derivation.toString();
-            turnString = turnString.replace(/ +/g, ' ');
-            turnString = this._sentenceGenerator.postprocess(turnString, userState);
-            mergeConstants(newConstants, turnString);
+            // set the turn of the agent
+            let state = derivation.value.state;
+            state = state.optimize();
+            assert(state !== null); // not-null even after optimize
+            this._stateValidator.validateAgent(state);
+            const utterance = this._postprocessSentence(derivation, state);
 
-            [agent, user] = turnString.split(' <sep> ');
-            assert(agent);
-            assert(user);
+            const dlg = derivation.context.priv;
+            assert(dlg instanceof PartialDialogue);
+            const prediction = this._target.computePrediction(dlg.context, state, 'agent');
+            const target = prediction.prettyprint();
 
-            const agentPrediction = this._target.computePrediction(dlg.context, agentState, 'agent');
-            agent_target = agentPrediction.prettyprint();
+            agentTurns.push({
+                dlg,
+                context: derivation.value.context,
+                tags: derivation.value.tags,
+                utterance,
+                state,
+                target
+            });
+        });
+        return agentTurns;
+    }
 
-            const userPrediction = this._target.computePrediction(agentState, userState, 'user');
-            user_target = userPrediction.prettyprint();
-        }
+    _generateUser(continuations, agentTurns) {
+        this._userGenerator.generate(agentTurns, (depth, derivation) => {
+            // the derivation value for the user is directly the thingtalk user state
+            // (unlike the agent)
 
-        continuations.put(dlg, {
-            userState,
-            newTurn: {
-                context: dlg.context !== null ? dlg.context.prettyprint() : null,
-                agent,
-                agent_target,
-                user,
-                user_target
-            },
-            newConstants: {}
+            let state = derivation.value;
+            state = state.optimize();
+            assert(state !== null); // not-null even after optimize
+            this._stateValidator.validateUser(state);
+            const utterance = this._postprocessSentence(derivation, state);
+
+            const agentTurn = derivation.context.priv;
+            const dlg = agentTurn.dlg;
+            assert(dlg instanceof PartialDialogue);
+            assert(dlg === this._emptyDialogue || agentTurn.state);
+            const prediction = this._target.computePrediction(agentTurn.state, state, 'user');
+            const target = prediction.prettyprint();
+
+            continuations.put(dlg, {
+                userState: state,
+                newTurn: {
+                    context: dlg.context !== null ? dlg.context.prettyprint() : null,
+                    agent: agentTurn.utterance,
+                    agent_target: agentTurn.target,
+                    user: utterance,
+                    user_target: target,
+                },
+            });
         });
     }
 
-    _tryAddTurn(continuations, derivation) {
-        try {
-            this._addTurn(continuations, derivation);
-        } catch(e) {
-            console.error(derivation.toString());
-            console.error(derivation.value);
-            throw e;
+    async _continueOneDialogue(dlg, continuations) {
+        const ourContinuations = continuations.get(dlg);
+
+        if (ourContinuations.length === 0) {
+            // if we have no continuations, mark this dialog as complete
+            this._maybeAddCompleteDialog(dlg);
+        } else {
+            for (let { userState, newTurn } of ourContinuations) {
+                const newDialogue = new PartialDialogue(dlg.context, dlg.turns.concat([newTurn]),
+                    dlg.execState);
+
+                if (this._maybeAddPartialDialog(newDialogue)) {
+                    const [newContext, newExecState] = await this._dialogueAgent.execute(userState, dlg.execState);
+                    newDialogue.context = newContext;
+                    newDialogue.execState = newExecState;
+                }
+            }
         }
     }
 
@@ -264,39 +285,24 @@ class MinibatchDialogueGenerator {
         this._partialDialogs.reset();
 
         const continuations = new MultiMap;
-        if (this._minibatchSize === 1) {
-            assert(partials.length <= 1);
-            if (partials.length > 0) {
-                const derivation = this._sentenceGenerator.generateOne(partials[0]);
-                if (derivation !== undefined)
-                    this._tryAddTurn(continuations, derivation);
-            }
+
+        let agentTurns = [];
+        if (this._turnIdx > 1) { // turnIdx is 1-based because we incremented it already
+            agentTurns = this._generateAgent(partials);
         } else {
-            this._sentenceGenerator.generate(partials, (depth, derivation) => {
-                this._tryAddTurn(continuations, derivation);
-            });
+            agentTurns = [{
+                dlg: this._emptyDialogue,
+                context: null,
+                tags: [],
+                utterance: '',
+                state: null,
+                target: null,
+            }];
         }
+        this._generateUser(continuations, agentTurns);
 
-        for (let i = 0; i < partials.length; i++) {
-            const dlg = partials[i];
-            const ourContinuations = continuations.get(dlg);
-
-            if (ourContinuations.length === 0) {
-                // if we have no continuations, mark this dialog as complete
-                this._maybeAddCompleteDialog(dlg);
-            } else {
-                for (let { userState, newTurn, newConstants } of ourContinuations) {
-                    const newDialogue = new PartialDialogue(dlg.context, dlg.turns.concat([newTurn]),
-                        newConstants, dlg.execState);
-
-                    if (this._maybeAddPartialDialog(newDialogue)) {
-                        const [newContext, newExecState] = await this._dialogueAgent.execute(userState, dlg.execState);
-                        newDialogue.context = newContext;
-                        newDialogue.execState = newExecState;
-                    }
-                }
-            }
-        }
+        for (let dlg of partials)
+            await this._continueOneDialogue(dlg, continuations);
 
         const end = Date.now();
         if (this._debug) {
@@ -314,6 +320,47 @@ class MinibatchDialogueGenerator {
             for (let dialogue of this._completeDialogs[turnIdx])
                 yield dialogue;
         }
+    }
+}
+
+class StateValidator {
+    constructor(policyManifest, target) {
+        this._policyManifest = policyManifest;
+        this._policy = null;
+        this._target = target;
+    }
+
+    async load() {
+        if (!this._policyManifest)
+            return;
+        const buffer = await util.promisify(fs.readFile)(this._policyManifest, { encoding: 'utf8' });
+        this._policy = yaml.safeLoad(buffer);
+        this._policy.dialogueActs.user = new Set(this._policy.dialogueActs.user);
+        this._policy.dialogueActs.agent = new Set(this._policy.dialogueActs.agent);
+        this._policy.dialogueActs.withParam = new Set(this._policy.dialogueActs.withParam);
+    }
+
+    validateUser(state) {
+        this._target.validateState(state, 'user');
+
+        if (!this._policy)
+            return;
+        assert.strictEqual(state.policy, this._policy.name);
+        assert(this._policy.dialogueActs.user.has(state.dialogueAct));
+        // if and only if
+        assert((state.dialogueActParam !== null) === (this._policy.dialogueActs.withParam.has(state.dialogueAct)));
+    }
+
+    validateAgent(state) {
+        this._target.validateState(state, 'agent');
+
+        if (!this._policy)
+            return;
+        assert.strictEqual(state.policy, this._policy.name);
+        assert(this._policy.dialogueActs.user.has(state.dialogueAct));
+        assert(state.dialogueAct !== this._policy.terminalAct);
+        // if and only if
+        assert((state.dialogueActParam !== null) === (this._policy.dialogueActs.withParam.has(state.dialogueAct)));
     }
 }
 
@@ -336,7 +383,31 @@ class DialogueGenerator extends stream.Readable {
         if (options.maxConstants === undefined)
             options.maxConstants = 5;
         this._target = require('../languages/' + options.targetLanguage);
-        this._generator = new SentenceGenerator(options);
+        this._langPack = i18n.get(options.locale);
+
+        const agentOptions = {};
+        Object.assign(agentOptions, options);
+        agentOptions.rootSymbol = '$agent';
+        agentOptions.contextInitializer = (partialDialogue, functionTable) => {
+            const tagger = functionTable.get('context');
+            return tagger(partialDialogue.context);
+        };
+        this._agentGenerator = new SentenceGenerator(agentOptions);
+
+        const userOptions = {};
+        Object.assign(userOptions, options);
+        userOptions.rootSymbol = '$user';
+        userOptions.contextInitializer = (agentTurn, functionTable) => {
+            if (agentTurn.context === null) {
+                // first turn
+                const tagger = functionTable.get('context');
+                return tagger(null);
+            } else {
+                return [agentTurn.tags, agentTurn.context];
+            }
+        };
+        this._userGenerator = new SentenceGenerator(userOptions);
+        this._stateValidator = new StateValidator(options.policyFile, this._target);
 
         this._initialized = false;
         this._dialogueAgent = this._target.createSimulator(options);
@@ -345,14 +416,18 @@ class DialogueGenerator extends stream.Readable {
 
     async _generateMinibatch() {
         if (!this._initialized) {
-            await this._generator.initialize();
+            await this._userGenerator.initialize();
+            await this._agentGenerator.initialize();
+            await this._stateValidator.load();
             this._initialized = true;
         }
 
         const start = Date.now();
         let counter = 0;
         try {
-            const generator = new MinibatchDialogueGenerator(this._generator, this._target, this._dialogueAgent, this._options, this._minibatchIdx++);
+            const generator = new MinibatchDialogueGenerator(this._agentGenerator,
+                this._userGenerator, this._target, this._langPack, this._dialogueAgent,
+                this._stateValidator, this._options, this._minibatchIdx++);
 
             for (let turn = 0; turn < this._options.maxTurns; turn++)
                 await generator.nextTurn();
