+++ lib/sentence-generator/runtime.js
@@ -43,16 +43,28 @@ class Placeholder {
     }
 }
 
+/**
+ * A reference to a context.
+ *
+ * A context is an object that is passed as extra input to a semantic function
+ * to affect its behavior. Grammar rules are only applied between identical (===) contexts.
+ *
+ * The "context" in this definition roughly corresponds to a dialogue context
+ * (either a C: state, or a more general notion) but it need not be.
+ *
+ * "priv" is a value associated with the context that is only meaningful to the API caller
+ * (DialogueGenerator). "info" is a value associated with the context that is meaningful
+ * to the semantic function.
+ * See {@link Grammar#_initializeContexts} for a longer explanation.
+ */
 class Context {
-    constructor(dlg, info) {
-        this.dlg = dlg;
-        this.constants = dlg.constants;
-        this.value = dlg.context;
+    constructor(priv, info) {
+        this.priv = priv;
         this.info = info;
     }
 
     toString() {
-        return `CTX[${this.value.prettyprint()}]`;
+        return `CTX[${this.info}]`;
     }
 }
 
@@ -281,10 +293,6 @@ function checkConstants(result, topLevel) {
         } else {
             if (topLevel) {
                 let min = 0;
-                if (result.context !== null)
-                    min = result.context.constants[piece.symbol] || 0;
-                else
-                    min = 0;
                 if (piece.number !== min) {
                     bad = true;
                     return;
@@ -367,11 +375,12 @@ class Grammar extends events.EventEmitter {
         this._rules = [];
 
         this._contextTable = new Map;
-        this._contextTagger = null;
+        this._functionTable = new Map;
 
         this._rootSymbol = options.rootSymbol || '$root';
         this._rootIndex = this._internalDeclareSymbol(this._rootSymbol);
         assert(this._rootIndex === 0);
+        this._contextInitializer = options.contextInitializer;
 
         // map constant tokens (QUOTED_STRING, NUMBER, ...) to the non-terms where they are used (constant_String, ...)
         this._constantMap = new MultiMap;
@@ -407,14 +416,12 @@ class Grammar extends events.EventEmitter {
         return index;
     }
 
-    setContextTagger(fn) {
+    declareFunction(name, fn) {
         if (this._finalized)
-            throw new GenieTypeError(`Grammar was finalized, cannot set context tagger`);
-        if (!this._contextual)
-            throw new GenieTypeError(`Grammar is not contextual, cannot set context tagger`);
-        if (this._contextTagger !== null)
-            throw new GenieTypeError(`Context tagger already set`);
-        this._contextTagger = fn;
+            throw new GenieTypeError(`Grammar was finalized, cannot declare more functions`);
+        if (this._functionTable.has(name))
+            throw new GenieTypeError(`Function ${name} already declared`);
+        this._functionTable.set(name, fn);
     }
 
     declareContext(context, fn) {
@@ -439,6 +446,9 @@ class Grammar extends events.EventEmitter {
             throw new GenieTypeError(`Identifier ${symbol} cannot be both a context and a non-terminal`);
         if (this.hasSymbol(symbol))
             return;
+        // ignore $user when loading the agent templates and vice-versa
+        if (symbol.startsWith('$') && symbol !== this._rootSymbol)
+            return;
 
         this._internalDeclareSymbol(symbol);
     }
@@ -458,6 +468,9 @@ class Grammar extends events.EventEmitter {
     addConstants(symbol, token, type, attributes = {}) {
         if (this._finalized)
             throw new GenieTypeError(`Grammar was finalized, cannot add more rules`);
+        // ignore $user when loading the agent templates and vice-versa
+        if (symbol.startsWith('$') && symbol !== this._rootSymbol)
+            return;
 
         const symbolId = this._lookupNonTerminal(symbol);
         this._constantMap.put(token, symbolId);
@@ -472,12 +485,15 @@ class Grammar extends events.EventEmitter {
     addRule(symbol, expansion, combiner, attributes) {
         if (this._finalized)
             throw new GenieTypeError(`Grammar was finalized, cannot add more rules`);
+        // ignore $user when loading the agent templates and vice-versa
+        if (symbol.startsWith('$') && symbol !== this._rootSymbol)
+            return;
         this._addRuleInternal(this._lookupNonTerminal(symbol), expansion, combiner, attributes);
     }
 
     _typecheck() {
-        if (this._contextual && !this._contextTagger)
-            throw new GenieTypeError(`Missing context tagger function for contextual grammar`);
+        if (this._contextual && !this._functionTable.has('context'))
+            throw new GenieTypeError(`Missing "context" function for contextual grammar`);
 
         for (let nonTermIndex = 0; nonTermIndex < this._nonTermList.length; nonTermIndex++) {
             const nonTerm = this._nonTermList[nonTermIndex];
@@ -688,33 +704,6 @@ class Grammar extends events.EventEmitter {
             }
         }
         return [estimate, ruleEstimates];
-    }
-
-    _initializeContexts(partialDialogues, charts, depth) {
-        for (let dlg of partialDialogues) {
-            try {
-                const result = this._contextTagger(dlg.context);
-                if (result !== null) {
-                    const [tags, info] = result;
-                    const ctx = new Context(dlg, info);
-                    for (let tag of tags) {
-                        const index = this._contextTable.get(tag);
-                        assert (index !== undefined, `Invalid context tag ${tag}`);
-                        charts[depth][index].push(ctx);
-                    }
-                }
-            } catch(e) {
-                console.error(dlg);
-                throw e;
-            }
-        }
-
-        if (this._options.debug) {
-            for (let index of this._contextTable.values()) {
-                if (charts[depth][index].length > 0)
-                    console.log(`stats: size(charts[${depth}][${this._nonTermList[index]}]) = ${charts[depth][index].length}`);
-            }
-        }
     }
 
     _enableAllRules() {
@@ -937,12 +926,68 @@ class Grammar extends events.EventEmitter {
     }
 
     /**
+     * Convert the context inputs into something suitable to pass to the semantic functions.
+     *
+     * `contextInputs` is an array of arbitrary objects, each corresponding to one possible
+     * context to use. Each object will be passed to the `contextInitializer` function that
+     * was passed when constructing the grammar.
+     *
+     * The contextInitializer returns a tuple with the "context tags" (a list of grammar
+     * non-terminals declared in a "context" statement) and the "context info", which is actually
+     * passed to the semantic function.
+     * The input to the context initializer and the info are stored in a Context object,
+     * which is then set as the .context property of the derivation.
+     *
+     * When generating agent sentences, contextInputs is an array of PartialDialogues,
+     * each having .context which is the C: state of the dialogue.
+     * We call the "context" function to compute the tags and the info.
+     *
+     * When generating user sentences, contextInputs is an array of agent turns. The turns
+     * already contain the tags and the context info (computed by the semantic function
+     * of the agent derivation).
+     *
+     * This fairly convoluted design allows two things:
+     * - "Grammar" (this class) has no notion of user or agent, all it knows is to put the right
+     *   stuff in the right non-terminals and generate.
+     * - The templates can define arbitrary tags for the context, and can define arbitrary functions
+     *   to tag the C: state context (the result of simulation).
+     * - The templates can pass arbitrary information from the agent turns to the user turns,
+     *   including information that is not representable in a ThingTalk dialogue state.
+     */
+    _initializeContexts(contextInputs, charts, depth) {
+        for (let ctxIn of contextInputs) {
+            try {
+                const result = this._contextInitializer(ctxIn, this._functionTable);
+                if (result !== null) {
+                    const [tags, info] = result;
+                    const ctx = new Context(ctxIn, info);
+                    for (let tag of tags) {
+                        const index = this._contextTable.get(tag);
+                        assert (index !== undefined, `Invalid context tag ${tag}`);
+                        charts[depth][index].push(ctx);
+                    }
+                }
+            } catch(e) {
+                console.error(ctxIn);
+                throw e;
+            }
+        }
+
+        if (this._options.debug) {
+            for (let index of this._contextTable.values()) {
+                if (charts[depth][index].length > 0)
+                    console.log(`stats: size(charts[${depth}][${this._nonTermList[index]}]) = ${charts[depth][index].length}`);
+            }
+        }
+    }
+
+    /**
      * Generate a batch of sentences (or dialogue turns), given the batch of contexts.
      *
      * This method is optimized for batch generation, and will cache intermediate generations
      * that do not depend on the context between subsequent calls.
      */
-    generate(contexts, callback) {
+    generate(contextInputs, callback) {
         this.finalize();
 
         // enable all rules (in case we called generateOne before)
@@ -1005,7 +1050,7 @@ class Grammar extends events.EventEmitter {
             assert(charts[depth].length === this._nonTermList.length);
 
             if (this._contextual && depth === 0)
-                this._initializeContexts(contexts, charts, depth);
+                this._initializeContexts(contextInputs, charts, depth);
 
             // compute estimates of how many things we will produce at this depth
             let [estimatedTotal, estimatedPerRule] = this._estimateDepthSize(charts, depth, firstGeneration);
