+++ package.json
@@ -23,9 +23,10 @@
     "en-pos": "^1.0.16",
     "errorhandler": "^1.5.1",
     "express": "^4.17.1",
-    "ip": "^1.1.5",
-    "matrix-js-sdk": "0.9.2",
     "flex-js": "^1.0.5",
+    "ip": "^1.1.5",
+    "js-yaml": "^3.14.0",
+    "matrix-js-sdk": "0.9.2",
     "morgan": "^1.9.1",
     "node-gettext": "^2.0.0",
     "q": "^1.5.0",
